Rename firestore-helpers for consistency.
