+++ server/functions/test/homestead/update-membership-on-creation.test.ts
@@ -1,14 +1,14 @@
 import { AdminFirestore, DocumentReference } from '@test-helpers/types';
 import { Collections } from '@test-helpers/constants';
 import { documentPath, generateUserId } from '@test-helpers/documents';
-
-import { waitForCloudFunctionExecution } from '../../helpers/wait';
 import {
   getAdminApp,
   setup,
   setUseRealProjectId,
   teardown,
-} from '../../../test-helpers/firestore-helpers';
+} from '@test-helpers/firestore';
+
+import { waitForCloudFunctionExecution } from '../../helpers/wait';
 
 const USER_ID = generateUserId();
