+++ server/rules/catch-all/delete.test.ts
@@ -1,9 +1,8 @@
 import { assertFails } from '@firebase/rules-unit-testing';
 import { Collections } from '@test-helpers/constants';
 import { generateId, generateUserId } from '@test-helpers/documents';
+import { setup, teardown } from '@test-helpers/firestore';
 import { Firestore } from '@test-helpers/types';
-
-import { setup, teardown } from '../../test-helpers/firestore-helpers';
 
 const COLLECTION = Collections.CatchAll;
 const DOC_ID = generateId();
