+++ server/rules/catch-all/update.test.ts
@@ -5,9 +5,8 @@ import {
   generateMockUpdateDocument,
   generateUserId,
 } from '@test-helpers/documents';
+import { setup, teardown } from '@test-helpers/firestore';
 import { Firestore } from '@test-helpers/types';
-
-import { setup, teardown } from '../../test-helpers/firestore-helpers';
 
 const COLLECTION = Collections.CatchAll;
 const DOC_ID = generateId();
