+++ server/rules/homestead/create.test.ts
@@ -7,13 +7,8 @@ import {
   generateMockUpdateDocument,
   generateUserId,
 } from '@test-helpers/documents';
+import { getAdminApp, setup, teardown } from '@test-helpers/firestore';
 import { Firestore } from '@test-helpers/types';
-
-import {
-  getAdminApp,
-  setup,
-  teardown,
-} from '../../test-helpers/firestore-helpers';
 
 const COLLECTION = Collections.Homesteads;
 const DOC_ID = generateId();
